refactor(amazonq): add BaseConnector Abstract class for common logic
