+++ packages/core/src/amazonq/webview/ui/apps/featureDevChatConnector.ts
@@ -3,21 +3,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { ChatItem, ChatItemAction, ChatItemType, FeedbackPayload } from '@aws/mynah-ui'
-import { ExtensionMessage } from '../commands'
-import { TabType, TabsStorage } from '../storages/tabsStorage'
-import { CodeReference } from './amazonqCommonsConnector'
-import { FollowUpGenerator } from '../followUps/generator'
+import { ChatItem, ChatItemType, FeedbackPayload } from '@aws/mynah-ui'
+import { TabType } from '../storages/tabsStorage'
 import { getActions } from '../diffTree/actions'
 import { DiffTreeFileInfo } from '../diffTree/types'
+import { BaseConnector, BaseConnectorProps } from './baseConnector'
 
-interface ChatPayload {
-    chatMessage: string
-}
-
-export interface ConnectorProps {
-    sendMessageToExtension: (message: ExtensionMessage) => void
-    onMessageReceived?: (tabID: string, messageData: any, needToShowAPIDocsTab: boolean) => void
+export interface ConnectorProps extends BaseConnectorProps {
     onAsyncEventProgress: (
         tabID: string,
         inProgress: boolean,
@@ -25,10 +17,7 @@ export interface ConnectorProps {
         messageId: string | undefined,
         enableStopAction: boolean
     ) => void
-    onChatAnswerReceived?: (tabID: string, message: ChatItem, messageData: any) => void
     sendFeedback?: (tabId: string, feedbackPayload: FeedbackPayload) => void | undefined
-    onError: (tabID: string, message: string, title: string) => void
-    onWarning: (tabID: string, message: string, title: string) => void
     onFileComponentUpdate: (
         tabID: string,
         filePaths: DiffTreeFileInfo[],
@@ -40,64 +29,28 @@ export interface ConnectorProps {
     onChatInputEnabled: (tabID: string, enabled: boolean) => void
     onUpdateAuthentication: (featureDevEnabled: boolean, authenticatingTabIDs: string[]) => void
     onNewTab: (tabType: TabType) => void
-    tabsStorage: TabsStorage
 }
 
-export class Connector {
-    private readonly sendMessageToExtension
-    private readonly onError
-    private readonly onWarning
+export class Connector extends BaseConnector {
     private readonly onFileComponentUpdate
-    private readonly onChatAnswerReceived
     private readonly onAsyncEventProgress
     private readonly updatePlaceholder
     private readonly chatInputEnabled
     private readonly onUpdateAuthentication
-    private readonly followUpGenerator: FollowUpGenerator
     private readonly onNewTab
 
+    override getTabType(): TabType {
+        return 'featuredev'
+    }
+
     constructor(props: ConnectorProps) {
-        this.sendMessageToExtension = props.sendMessageToExtension
-        this.onChatAnswerReceived = props.onChatAnswerReceived
-        this.onWarning = props.onWarning
+        super(props)
         this.onFileComponentUpdate = props.onFileComponentUpdate
-        this.onError = props.onError
         this.onAsyncEventProgress = props.onAsyncEventProgress
         this.updatePlaceholder = props.onUpdatePlaceholder
         this.chatInputEnabled = props.onChatInputEnabled
         this.onUpdateAuthentication = props.onUpdateAuthentication
-        this.followUpGenerator = new FollowUpGenerator()
         this.onNewTab = props.onNewTab
-    }
-
-    onCodeInsertToCursorPosition = (
-        tabID: string,
-        code?: string,
-        type?: 'selection' | 'block',
-        codeReference?: CodeReference[]
-    ): void => {
-        this.sendMessageToExtension({
-            tabID: tabID,
-            code,
-            command: 'insert_code_at_cursor_position',
-            codeReference,
-            tabType: 'featuredev',
-        })
-    }
-
-    onCopyCodeToClipboard = (
-        tabID: string,
-        code?: string,
-        type?: 'selection' | 'block',
-        codeReference?: CodeReference[]
-    ): void => {
-        this.sendMessageToExtension({
-            tabID: tabID,
-            code,
-            command: 'code_was_copied_to_clipboard',
-            codeReference,
-            tabType: 'featuredev',
-        })
     }
 
     onOpenDiff = (tabID: string, filePath: string, deleted: boolean, messageId?: string): void => {
@@ -107,7 +60,7 @@ export class Connector {
             filePath,
             deleted,
             messageId,
-            tabType: 'featuredev',
+            tabType: this.getTabType(),
         })
     }
     onFileActionClick = (tabID: string, messageId: string, filePath: string, actionName: string): void => {
@@ -117,28 +70,9 @@ export class Connector {
             messageId,
             filePath,
             actionName,
-            tabType: 'featuredev',
+            tabType: this.getTabType(),
         })
     }
-
-    followUpClicked = (tabID: string, followUp: ChatItemAction): void => {
-        this.sendMessageToExtension({
-            command: 'follow-up-was-clicked',
-            followUp,
-            tabID,
-            tabType: 'featuredev',
-        })
-    }
-
-    requestGenerativeAIAnswer = (tabID: string, payload: ChatPayload): Promise<any> =>
-        new Promise((resolve, reject) => {
-            this.sendMessageToExtension({
-                tabID: tabID,
-                command: 'chat-prompt',
-                chatMessage: payload.chatMessage,
-                tabType: 'featuredev',
-            })
-        })
 
     private processChatMessage = async (messageData: any): Promise<void> => {
         if (this.onChatAnswerReceived !== undefined) {
@@ -191,36 +125,6 @@ export class Connector {
         }
     }
 
-    private processAuthNeededException = async (messageData: any): Promise<void> => {
-        if (this.onChatAnswerReceived === undefined) {
-            return
-        }
-
-        this.onChatAnswerReceived(
-            messageData.tabID,
-            {
-                type: ChatItemType.ANSWER,
-                body: messageData.message,
-                followUp: undefined,
-                canBeVoted: false,
-            },
-            messageData
-        )
-
-        this.onChatAnswerReceived(
-            messageData.tabID,
-            {
-                type: ChatItemType.SYSTEM_PROMPT,
-                body: undefined,
-                followUp: this.followUpGenerator.generateAuthFollowUps('featuredev', messageData.authType),
-                canBeVoted: false,
-            },
-            messageData
-        )
-
-        return
-    }
-
     handleMessageReceive = async (messageData: any): Promise<void> => {
         if (messageData.type === 'updateFileComponent') {
             this.onFileComponentUpdate(
@@ -229,15 +133,6 @@ export class Connector {
                 messageData.deletedFiles,
                 messageData.messageId
             )
-            return
-        }
-        if (messageData.type === 'errorMessage') {
-            this.onError(messageData.tabID, messageData.message, messageData.title)
-            return
-        }
-
-        if (messageData.type === 'showInvalidTokenNotification') {
-            this.onWarning(messageData.tabID, messageData.message, messageData.title)
             return
         }
 
@@ -278,67 +173,21 @@ export class Connector {
             return
         }
 
-        if (messageData.type === 'authNeededException') {
-            await this.processAuthNeededException(messageData)
-            return
-        }
-
         if (messageData.type === 'openNewTabMessage') {
             this.onNewTab('featuredev')
             return
         }
-    }
 
-    onStopChatResponse = (tabID: string): void => {
-        this.sendMessageToExtension({
-            tabID: tabID,
-            command: 'stop-response',
-            tabType: 'featuredev',
-        })
-    }
-
-    onTabOpen = (tabID: string): void => {
-        this.sendMessageToExtension({
-            tabID,
-            command: 'new-tab-was-created',
-            tabType: 'featuredev',
-        })
-    }
-
-    onTabRemove = (tabID: string): void => {
-        this.sendMessageToExtension({
-            tabID: tabID,
-            command: 'tab-was-removed',
-            tabType: 'featuredev',
-        })
+        // For other message types, call the base class handleMessageReceive
+        await this.baseHandleMessageReceive(messageData)
     }
 
     sendFeedback = (tabId: string, feedbackPayload: FeedbackPayload): void | undefined => {
         this.sendMessageToExtension({
             command: 'chat-item-feedback',
             ...feedbackPayload,
-            tabType: 'featuredev',
+            tabType: this.getTabType(),
             tabID: tabId,
-        })
-    }
-
-    onChatItemVoted = (tabId: string, messageId: string, vote: string): void | undefined => {
-        this.sendMessageToExtension({
-            tabID: tabId,
-            messageId: messageId,
-            vote: vote,
-            command: 'chat-item-voted',
-            tabType: 'featuredev',
-        })
-    }
-
-    onResponseBodyLinkClick = (tabID: string, messageId: string, link: string): void => {
-        this.sendMessageToExtension({
-            command: 'response-body-link-click',
-            tabID,
-            messageId,
-            link,
-            tabType: 'featuredev',
         })
     }
 }
